backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
qdrant-client
sentence-transformers
python-multipart
python-dotenv
pydantic
pypdf2
python-docx
pandas
numpy
fastapi-cors
agno==2.0.3
ddgs
openai
redis
aioredis